+++ src/device.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <stddef.h>
+#include <stdint.h>
 #include "cursor.h"
 
 /**
@@ -22,11 +22,11 @@ heap
 typedef struct {
   cursor_t* mmap;
   // Exact byte length of block device.
-  size_t size;
+  uint64_t size;
   // This should equal ceil(device_size / tile_size).
   uint32_t tile_count;
 } device_t;
 
-device_t* device_create(void* mmap, size_t size);
+device_t* device_create(void* mmap, uint64_t size);
 
 void device_format(device_t* dev, uint8_t bucket_count_log2);
