+++ src/list.h
@@ -1,16 +1,16 @@
 #pragma once
 
-#include <stddef.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 #define LIST_DEF(name, type_t) \
   typedef struct { \
-    size_t cap; \
-    size_t len; \
+    uint64_t cap; \
+    uint64_t len; \
     type_t* elems; \
   } name##_t; \
   \
-  name##_t* name##_create_with_capacity(size_t init_cap); \
+  name##_t* name##_create_with_capacity(uint64_t init_cap); \
   \
   name##_t* name##_create(); \
   \
@@ -19,7 +19,7 @@
   type_t* name##_last_mut(name##_t* list);
 
 #define LIST(name, type_t) \
-  name##_t* name##_create_with_capacity(size_t init_cap) { \
+  name##_t* name##_create_with_capacity(uint64_t init_cap) { \
     name##_t* list = malloc(sizeof(name##_t)); \
     list->cap = init_cap; \
     list->len = 0; \
@@ -33,7 +33,7 @@
   \
   void name##_append(name##_t* list, type_t elem) { \
     if (list->len == list->cap) { \
-      size_t new_cap = list->cap * 2; \
+      uint64_t new_cap = list->cap * 2; \
       list->elems = realloc(list->elems, new_cap); \
       list->cap = new_cap; \
     } \
